api/requirements.txt
fastapi
uvicorn[standard]
gtfs-realtime-bindings
protobuf
psycopg2-binary
